split loaders into separate files
