+++ geoflow-core/src/main/kotlin/loading/AnalyzeInfo.kt
@@ -7,7 +7,7 @@ data class AnalyzeInfo(
     /** table name for the file/sub table */
     val tableName: String,
     /** delimiter for the source table, if needed */
-    val delimiter: Char = defaultDelimiter,
+    val delimiter: Char = DEFAULT_DELIMITER,
     /** flag denoting if the source table data is qualified, if needed */
     val qualified: Boolean = true,
     /** sub table if the source file has multiple tables in a single file, null if the file has a single dataset */
